BERT-BiLSTM-CRF/requirements.txt
transformers
torch
seqeval
pytorch-crf
matplotlib